usb: dwc3-msm: Reordering the expression in "if" check
Currently in dwc3_otg_start_peripheral there is a "if" to find out
if we are bound to udc or not. In this change reordering the expression
in "if" check to avoid null pointer deference.

Flow: dwc3_otg_start_peripheral -> we check for dwc->gadget->udc->driver
but it is already null which is why it will through error for null
pointer deference so to avoid this doing short-circuiting.

Adding a check for dwc->gadget before dereference.

Change-Id: I69521c84b681660d6be6ecedae8aaec235534c4d
Signed-off-by: Uttkarsh Aggarwal <quic_uaggarwa@quicinc.com>