spi: spi-msm-geni: spi hibernation support for gsi mode
While entering hibernation in GSI mode, if the device is
already in runtime suspend mode, gsi_pause is not invoked
with the deep sleep init flag.

Perform runtime resume forcefully if the device is already
in runtime suspend while entering hibernation. And invoke
runtime suspend forcefully to get the impact of the deep
sleep init flag in GSI pause.

Change-Id: I418f65200d66536fff53a7170e3f942d7d430ec8
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>